# docs/api/openapi.core.v1.yaml
openapi: 3.1.0
info:
  title: DELMOS Core First Slice API
  version: 0.1.0-draft
  description: >-
    Target contract for the first Core role and approval slice. Not implemented.
    Session establishment and external integrations are outside this fragment.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /projects/{project_id}/work-products:
    post:
      operationId: createWorkProduct
      summary: Create a draft Work Product with its first immutable revision
      x-required-permission: wp.create
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkProduct'
      responses:
        '201':
          description: Draft created with server-generated revision and payload hash
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkProductState'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
  /projects/{project_id}/work-products/{work_product_id}/revisions:
    post:
      operationId: reviseWorkProduct
      summary: Write a new immutable draft revision
      x-required-permission: wp.edit
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/WorkProductId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviseWorkProduct'
      responses:
        '201':
          description: New revision created without modifying the prior revision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkProductState'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Invariant' }
...
  /projects/{project_id}/work-products/{work_product_id}/similar:
    get:
      operationId: findSimilarWorkProducts
      summary: Hybrid semantic search for near-duplicate requirements/tests (VEC-01..03)
      x-required-permission: wp.read
      description: >-
        Compares the work product's current revision embedding against other
        embedded revisions in the same project via pgvector cosine distance.
        Only vectors of the current embedding model are matched; permission
        scope is applied as a project_id SQL filter, not post-filtering.
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: work_product_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Similar work products ordered by descending cosine similarity
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SimilarWorkProduct' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/Hidden' }
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: delmos_session
      description: Server-side HttpOnly session; its creation is outside this fragment.
  parameters:
...
    ProjectId:
      in: path
      name: project_id
      required: true
      schema: { type: string, format: uuid }
...
    WorkProductId:
      in: path
      name: work_product_id
      required: true
      schema: { type: string, format: uuid }
...
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string, minLength: 1 }
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: No valid session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Missing permission for a visible action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Hidden:
      description: Resource is missing or not visible in this scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Stale row version or idempotency key reused with different payload
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Invariant:
      description: Validation, SoD, quorum or lifecycle rule failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
...
    CreateWorkProduct:
      type: object
      additionalProperties: false
      required: [code, type, title, body]
      properties:
        code: { type: string, minLength: 1 }
        type: { const: requirement }
        title: { type: string, minLength: 1 }
        body: { type: string }
        metadata: { type: object }
...
    SimilarWorkProduct:
      type: object
      required: [work_product_id, code, title, status, cosine_similarity]
      properties:
        work_product_id: { type: string, format: uuid }
        code: { type: string }
        title: { type: string }
        status: { type: string }
        cosine_similarity: { type: number, format: double, minimum: -1, maximum: 1 }
...
    ReviseWorkProduct:
      type: object
      additionalProperties: false
      required: [expected_row_version, title, body]
      properties:
        expected_row_version: { type: integer, minimum: 1 }
        title: { type: string, minLength: 1 }
        body: { type: string }
        metadata: { type: object }
...
    WorkProductState:
      type: object
      required: [work_product_id, revision_id, payload_hash, row_version, status]
      properties:
        work_product_id: { type: string, format: uuid }
        revision_id: { type: string, format: uuid }
        payload_hash: { type: string, pattern: '^[0-9a-f]{64}$' }
        row_version: { type: integer, minimum: 1 }
        status: { type: string, enum: [draft, in_review, approved, obsolete] }
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  issue: { type: string }